Root/PPT_Agent/requirements.txt
python-pptx
pydantic
openai
python-dotenv
matplotlib
seaborn
plotly
vizro
pygwalker
statsmodels
scikit-learn
scipy
diagrams
qbstyles
html2image
kaleido
google-generativeai
pandas
numpy
